// src/enums/PokerHandEnum.ts
export enum PokerHandEnum {
  STRAIGHT_FLUSH_ID = "straight-flush",
  FOUR_OF_A_KIND_ID = "four-of-a-kind",
  FULL_HOUSE_ID = "full-house",
  FLUSH_ID = "flush",
  STRAIGHT_ID = "straight",
  THREE_OF_A_KIND_ID = "three-of-a-kind",
  TWO_PAIRS_ID = "two-pairs",
  PAIR_ID = "pair",
  HIGH_CARD_ID = "high-card",
}